V 1.0.1

+++ lib/app/domain/entities/product_entity/get_all_product_entity.dart
@@ -1,11 +1,11 @@
 class GetAllProductEntity {
   int? id;
   String? name;
-  int? totalQuantity;
-  int? remainingQuantity;
+  double? totalQuantity;
+  double? remainingQuantity;
   double? latestPricePerKg;
-  int? totalPrice;
-  int? remainingPrice;
+  double? totalPrice;
+  double? remainingPrice;
   double? averageUnitPrice;
 
   GetAllProductEntity({
